requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
attrs==22.1.0
colorama==0.4.6
exceptiongroup==1.0.4
importlib-metadata==5.1.0
iniconfig==1.1.1
packaging==22.0
pluggy==1.0.0
pytest==7.2.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.4.0
zipp==3.11.0
